[ENH] Separate tests files to prevent non-execution of some of them
