+++ tests/phpunit/config.php
@@ -156,177 +156,3 @@ class Hm_Test_User_File_Config extends TestCase {
         unset($this->config);
     }
 }
-
-/**
- * tests for Hm_Test_Site_File_Config
- */
-class Hm_Test_Site_File_Config extends TestCase {
-
-    public $config;
-    public function setUp(): void {
-        require 'bootstrap.php';
-        $mock_config = new Hm_Mock_Config();
-        $this->config = new Hm_User_Config_File($mock_config);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_get_modules() {
-        $config = new Hm_Site_Config_File(merge_config_files(APP_PATH.'tests/phpunit/data'));
-        $this->assertFalse($config->get_modules());
-        $config->set('modules', 'asdf');
-        $this->assertEquals(array('asdf'), $config->get_modules());
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_site_load() {
-        $config = new Hm_Site_Config_File(merge_config_files(APP_PATH.'tests/phpunit/data'));
-        $this->assertEquals(array('version' => VERSION, 'foo' => 'bar', 'default_setting_foo' => 'bar'), $config->dump());
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_get_user_defaults() {
-        $config = new Hm_Site_Config_File(merge_config_files(APP_PATH.'tests/phpunit/data'));
-        $this->assertEquals(array('version' => VERSION, 'foo' => 'bar', 'default_setting_foo' => 'bar'), $config->dump());
-    }
-    public function tearDown(): void {
-        unset($this->config);
-    }
-}
-
-/**
- * tests for Hm_User_Config_DB
- */
-class Hm_Test_User_Config_DB extends TestCase {
-
-    public $config;
-    public function setUp(): void {
-        require 'bootstrap.php';
-        $mock_config = new Hm_Mock_Config();
-        $this->config = new Hm_User_Config_File($mock_config);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_connect() {
-        $site_config = new Hm_Mock_Config();
-        setup_db($site_config);
-        $user_config = new Hm_User_Config_DB($site_config);
-        $this->assertTrue($user_config->connect());
-    }
-
-    /*public function test_db_load() {
-        $site_config = new Hm_Mock_Config();
-        setup_db($site_config);
-        $user_config = new Hm_User_Config_DB($site_config);
-        $this->assertEquals(array('version' => VERSION), $user_config->dump());
-        $user_config->load('testuser', 'testkey');
-        $this->assertEquals(array('version' => VERSION, 'foo' => 'bar'), $user_config->dump());
-        $user_config->reload(array());
-        $user_config->load('testuser', 'testpass');
-        $this->assertEquals(array(), $user_config->dump());
-        $user_config->reload(array());
-        $user_config->load(uniqid(), 'blah');
-        $this->assertEquals(array(), $user_config->dump());
-        $site_config->set('auth_type', 'IMAP');
-        $site_config->set('single_server_mode', true);
-        $user_config = new Hm_User_Config_DB($site_config);
-        $user_config->load('testuser', 'testkey');
-        $this->assertTrue(array_key_exists('version', $user_config->dump()));
-    }*/
-
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_reload() {
-        $site_config = new Hm_Mock_Config();
-        setup_db($site_config);
-        $user_config = new Hm_User_Config_DB($site_config);
-        $user_config->reload(array('foo' => 'bar'));
-        $this->assertEquals(array('foo' => 'bar'), $user_config->dump());
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_save() {
-        $site_config = new Hm_Mock_Config();
-        $user_config = new Hm_User_Config_DB($site_config);
-        $user_config->reload(array('foo' => 'bar'));
-        $this->assertFalse($user_config->save('testuser', 'testkey'));
-
-        $site_config = new Hm_Mock_Config();
-        setup_db($site_config);
-        $user_config = new Hm_User_Config_DB($site_config);
-        $user_config->reload(array('foo' => 'bar'));
-        $this->assertTrue($user_config->save('testuser', 'testkey'));
-        $this->assertEquals(1, $user_config->save(uniqid(), 'testkey'));
-        //$this->assertFalse($user_config->save(NULL, 'blah'));
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_set() {
-        $site_config = new Hm_Mock_Config();
-        setup_db($site_config);
-        $user_config = new Hm_User_Config_DB($site_config);
-        $user_config->set('foo',  'bar');
-        $this->assertEquals('bar', $user_config->get('foo'));
-        $site_config->set('auth_type', 'IMAP');
-        $site_config->set('single_server_mode', true);
-        $user_config = new Hm_User_Config_DB($site_config);
-        $user_config->load('foo', 'foo');
-        $user_config->set('foo',  'bar');
-        $this->assertEquals('bar', $user_config->get('foo'));
-    }
-    public function tearDown(): void {
-        unset($this->config);
-    }
-}
-
-class Hm_Test_User_Config_Functions extends TestCase {
-
-    public $config;
-    public function setUp(): void {
-        require 'bootstrap.php';
-        $mock_config = new Hm_Mock_Config();
-        $this->config = new Hm_User_Config_File($mock_config);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_load_user_config_object() {
-        /* TODO assertions */
-        $mock_config = new Hm_Mock_Config();
-        load_user_config_object($mock_config);
-        $this->assertEquals('Hm_User_Config_File', get_class(load_user_config_object($mock_config)));
-        $mock_config->set('user_config_type', 'DB');
-        $this->assertEquals('Hm_User_Config_DB', get_class(load_user_config_object($mock_config)));
-        $mock_config->set('user_config_type', 'custom:Hm_Mock_Config');
-        $this->assertEquals('Hm_Mock_Config', get_class(load_user_config_object($mock_config)));
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_crypt_state() {
-        $site_config = new Hm_Mock_Config();
-        $this->assertTrue(crypt_state($site_config));
-        $site_config->set('auth_type', 'IMAP');
-        $site_config->set('single_server_mode', true);
-        $this->assertFalse(crypt_state($site_config));
-    }
-
-    public function tearDown(): void {
-        unset($this->config);
-    }
-}
