+++ tests/phpunit/session.php
@@ -268,349 +268,3 @@ class Hm_Test_PHP_Session extends TestCase {
         $session->destroy($request);
     }
 }
-
-/**
- * tests for Hm_Redis_Session
- */
-class Hm_Test_Redis_Session extends TestCase {
-
-    public $config;
-    public function setUp(): void {
-        require 'bootstrap.php';
-        ini_set('session.use_cookies', '0');
-        session_cache_limiter('');
-        $this->config = new Hm_Mock_Config();
-        $this->config->set('redis_server', 'asdf');
-        $this->config->set('redis_port', 10);
-        $this->config->set('enable_redis', true);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_redis_connect() {
-        $session = new Hm_Redis_Session($this->config, 'Hm_Auth_DB');
-        $session->connect();
-        $this->assertEquals('Hm_Redis', get_class($session->conn));
-    }
-}
-
-/**
- * tests for Hm_Memcached_Session
- */
-class Hm_Test_Memcached_Session extends TestCase {
-
-    public $config;
-    public function setUp(): void {
-        require 'bootstrap.php';
-        ini_set('session.use_cookies', '0');
-        session_cache_limiter('');
-        $this->config = new Hm_Mock_Config();
-        $this->config->set('memcached_server', 'asdf');
-        $this->config->set('memcached_port', 10);
-        $this->config->set('enable_memcached', true);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_memcached_connect() {
-        $session = new Hm_Memcached_Session($this->config, 'Hm_Auth_DB');
-        $session->connect();
-        $this->assertEquals('Hm_Memcached', get_class($session->conn));
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_memcached_start_existing() {
-        $session = new Hm_Memcached_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->loaded = true;
-        $session->start($request);
-        $session->set('foo', 'bar');
-        $session->save_data();
-        $session->start_existing($session->session_key);
-        $this->assertEquals('bar', $session->get('foo'));
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_memcached_start() {
-        $session = new Hm_Memcached_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->loaded = true;
-        $session->start($request);
-        $this->assertTrue($session->is_active());
-        $session->destroy($request);
-
-        $request->cookie['hm_session'] = 'test';
-        $session->loaded = false;
-        $session->start($request);
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-
-        $request->cookie = array();
-        $session->loaded = false;
-        $session->start($request);
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-
-        Hm_Mock_Memcached::$set_failure = true;
-        $session = new Hm_Memcached_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->loaded = true;
-        $session->start($request);
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_memcached_end() {
-        $session = new Hm_Memcached_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->connect();
-        $session->end();
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-
-        $session = new Hm_Memcached_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->loaded = true;
-        $session->start($request);
-        $session->end();
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_memcached_close_early() {
-        $session = new Hm_Memcached_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->connect();
-        $session->close_early();
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-    }
-}
-
-/**
- * tests for Hm_DB_Session
- */
-class Hm_Test_DB_Session extends TestCase {
-
-    public $config;
-    public function setUp(): void {
-        require 'bootstrap.php';
-        ini_set('session.use_cookies', '0');
-        session_cache_limiter('');
-        $this->config = new Hm_Mock_Config();
-        setup_db($this->config);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_connect() {
-        $_POST['user'] = 'unittestusers';
-        $_POST['pass'] = 'unittestpass';
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $this->assertTrue($session->connect());
-        $session->destroy($request);
-        $config = new Hm_Mock_Config();
-        $session = new Hm_DB_Session($config, 'Hm_Auth_DB');
-        $this->assertFalse($session->connect());
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_end() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->end();
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_start() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->loaded = true;
-        $session->start($request);
-        $this->assertTrue($session->is_active());
-        $session->destroy($request);
-
-        $request->cookie['hm_session'] = 'test';
-        $session->loaded = false;
-        $session->start($request);
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-
-        $request->cookie = array();
-        $session->loaded = false;
-        $session->start($request);
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_start_existing_session() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->loaded = true;
-        $session->start($request);
-        $this->assertTrue($session->is_active());
-        $key = $session->session_key;
-        $session->end();
-        $session->start_existing($key);
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_close_early() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->close_early();
-        $this->assertFalse($session->is_active());
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_db_save_data() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $session->loaded = true;
-        $session->start($request);
-        $this->assertEquals(1, $session->save_data());
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_plaintext() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $this->assertEquals(array('data'), ($session->plaintext($session->ciphertext(array('data')))));
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_get_key() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $request->cookie['hm_id'] = 'test';
-        $session->get_key($request);
-        $this->assertEquals('test', $session->enc_key);
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_secure_cookie() {
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_DB');
-        $request = new Hm_Mock_Request('HTTP');
-        $request->tls = true;
-        $request->path = 'test';
-        $this->assertTrue($session->secure_cookie($request, 'name', 'value'));
-        $this->assertTrue($session->secure_cookie($request, 'name', 'value', '/', 'http://localhost:123'));
-        $session->destroy($request);
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_insert_session_row() {
-        $request = new Hm_Mock_Request('HTTP');
-        $session = new Hm_DB_Session($this->config, 'Hm_Auth_None');
-        $this->assertFalse($session->insert_session_row());
-        $session->connect();
-        $this->assertEquals(1, $session->insert_session_row());
-        $session->destroy($request);
-
-        $config = new Hm_Mock_Config();
-        $session = new Hm_DB_Session($config, 'Hm_Auth_DB');
-        $session->connect();
-        $this->assertFalse($session->insert_session_row());
-        $session->destroy($request);
-
-    }
-    public function tearDown(): void {
-        unset($this->config);
-    }
-}
-
-class Hm_Test_Session_Functions extends TestCase {
-
-    public $config;
-    public function setUp(): void {
-        require 'bootstrap.php';
-        $this->config = new Hm_Mock_Config();
-        require APP_PATH.'modules/site/lib.php';
-    }
-    /**
-     * @preserveGlobalState disabled
-     * @runInSeparateProcess
-     */
-    public function test_setup_session() {
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Hm_PHP_Session', get_class(($setup->setup_session())));
-
-        $this->config->set('session_type', 'DB');
-        $this->config->set('auth_type', 'DB');
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Hm_DB_Session', get_class(($setup->setup_session())));
-
-        $this->config->mods[] = 'dynamic_login';
-        $this->config->set('auth_type', 'dynamic');
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Hm_DB_Session', get_class(($setup->setup_session())));
-
-        $this->config->set('session_type', 'asdf');
-        $this->config->set('auth_type', 'DB');
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Hm_PHP_Session', get_class(($setup->setup_session())));
-
-        $this->config->set('session_type', 'custom');
-        $this->config->set('auth_type', 'custom');
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Custom_Session', get_class(($setup->setup_session())));
-
-        $this->config->set('session_type', 'REDIS');
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Hm_Redis_Session', get_class(($setup->setup_session())));
-
-        $this->config->set('session_type', 'MEM');
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Hm_Memcached_Session', get_class(($setup->setup_session())));
-
-        Hm_Functions::$exists = false;
-        $this->config->set('session_type', 'PHP');
-        $this->config->set('auth_type', 'asdf');
-        $setup = new Hm_Session_Setup($this->config);
-        $this->assertEquals('Hm_PHP_Session', get_class(($setup->setup_session())));
-
-    }
-    public function tearDown(): void {
-        unset($this->config);
-    }
-}
